# go/model_address.go
/*
 * ModelDB API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Address struct {

	Id int64 `json:"id,omitempty"`

	Street string `json:"street,omitempty"`

	Number int32 `json:"number,omitempty"`

	City string `json:"city,omitempty"`

	Postalcode int32 `json:"postalcode,omitempty"`

	Country string `json:"country,omitempty"`
}