# go/model_contract.go
/*
 * ModelDB API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Contract struct {

	Id int64 `json:"id,omitempty"`

	Type_ string `json:"type,omitempty"`

	Version string `json:"version,omitempty"`

	Date string `json:"date,omitempty"`

	Active bool `json:"active,omitempty"`

	File string `json:"file,omitempty"`
}