# go/model_idea.go
/*
 * ModelDB API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Idea struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Content string `json:"content,omitempty"`
}