# go/model_model.go
/*
 * ModelDB API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Model struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Surname string `json:"surname,omitempty"`

	Address *Address `json:"address,omitempty"`

	Birthdate string `json:"birthdate,omitempty"`

	Contract []Contract `json:"contract,omitempty"`

	Tags []Tag `json:"tags,omitempty"`
}